==> 99_spec_adjsustment.ipynb <==
import pandas as pd

spec1 = pd.read_pickle('data/specialities/specialities.pkl')
spec2 = pd.read_csv('data/specialities/manual_specialities.csv', sep=';')

map1 = pd.read_pickle('data/specialities/speciality_mapping.pkl')
map2 = pd.read_csv('data/specialities/manual_speciality_mapping.csv', sep=';')

spec = pd.concat([spec1, spec2], ignore_index=True)
map = pd.concat([map1, map2], ignore_index=True)

spec_map = pd.merge(spec, map, left_on='speciality', right_on='German', how='left')


spec_map = spec_map.drop(columns=['German'])

spec_map.to_csv('data/specialities/all_specialities_mapping_de_en.csv', sep=';', index=False)
